% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Legend.R
\docType{data}
\name{Legend}
\alias{Legend}
\title{This dataset contains CCSR codes that are associated with over all ICD-10 codes}
\format{
A data frame with 543 rows and 3 variables:
 \describe{
   \item{CCSRCat}{CCSR category corresponding to ICD-10 code}
   \item{CCSRD}{Text description of CCSR category}
   \item{codecolvec}{A regular expression listing all the ICD-10 codes for the CCSR code}
   }
}
\source{
\url{https://hcup-us.ahrq.gov/toolssoftware/ccsr/dxccsr.jsp}
}
\usage{
Legend
}
\description{
This dataset contains CCSR codes that are associated with over all ICD-10 codes
}
\examples{
 data(Legend)
 head(Legend)
}
\keyword{datasets}
